refactor(ci): extract shared script helpers into ci_common
(cherry picked from commit a0b5004bf8a2e09f8cc3a55c9cb0b632f3693cf2)
